+++ .gitignore
@@ -197,7 +197,6 @@ nbs/
 
 old_stuff/
 
-src/facts_experiment_builder/adapters/experiment_adapters.py
 src/facts_experiment_builder/core/total_checker.py
 src/facts_experiment_builder/core/workspace/
 src/facts_experiment_builder/experiment_recipes.yml
