+++ tests/unit/test_setup_experiment.py
@@ -5,7 +5,7 @@ from facts_experiment_builder.application.setup_experiment import (
     prepare_experiment_setup,
     finalize_experiment_setup,
 )
-from facts_experiment_builder.adapters.experiment_adapters import (
+from facts_experiment_builder.application.experiment_helpers import (
     hydrate_experiment,
 )
 from facts_experiment_builder.infra.write_experiment_metadata import format_module_value
